Extract queries within explorer
